requirements.txt
numpy
pillow
scipy
scikit-image